+++ compiler-tests/src/test/data/box/dependencygraph/extensions/ScopedExtensionsWithoutFactoriesShouldBeCached.kt
@@ -1,0 +1,24 @@
+abstract class LoggedInScope
+
+@GraphExtension
+@SingleIn(LoggedInScope::class)
+interface LoggedInGraph {
+  val int: Int
+}
+
+@DependencyGraph
+interface AppGraph {
+  @Provides fun provideInt(): Int = 3
+
+  fun loggedInGraph(): LoggedInGraph
+  val loggedInGraphProp: LoggedInGraph
+}
+
+fun box(): String {
+  val graph = createGraph<AppGraph>()
+  val loggedInGraph = graph.loggedInGraph()
+  assertEquals(3, loggedInGraph.int)
+  assertSame(loggedInGraph, graph.loggedInGraph())
+  assertSame(graph.loggedInGraphProp, graph.loggedInGraph())
+  return "OK"
+}