+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -168,6 +168,28 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     }
 
     @Nested
+    @TestMetadata("compiler-tests/src/test/data/diagnostic/dependencygraph/extensions")
+    @TestDataPath("$PROJECT_ROOT")
+    public class Extensions {
+      @Test
+      public void testAllFilesPresentInExtensions() {
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/diagnostic/dependencygraph/extensions"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
+      }
+
+      @Test
+      @TestMetadata("ExtensionAccessorsCannotHaveParameters.kt")
+      public void testExtensionAccessorsCannotHaveParameters() {
+        runTest("compiler-tests/src/test/data/diagnostic/dependencygraph/extensions/ExtensionAccessorsCannotHaveParameters.kt");
+      }
+
+      @Test
+      @TestMetadata("ExtensionsWithFactoriesMustUseThem.kt")
+      public void testExtensionsWithFactoriesMustUseThem() {
+        runTest("compiler-tests/src/test/data/diagnostic/dependencygraph/extensions/ExtensionsWithFactoriesMustUseThem.kt");
+      }
+    }
+
+    @Nested
     @TestMetadata("compiler-tests/src/test/data/diagnostic/dependencygraph/leniency")
     @TestDataPath("$PROJECT_ROOT")
     public class Leniency {